+++ examples/foxes/timeseries/run.py
@@ -95,7 +95,7 @@ def run_foxes(args):
     P = o.calc_mean_farm_power()
     print(f"\nFarm power        : {P/1000:.1f} MW")
     print(f"Farm ambient power: {P0/1000:.1f} MW")
-    print(f"Farm efficiency   : {o.calc_farm_efficiency():.2f}")
+    print(f"Farm efficiency   : {o.calc_farm_efficiency()*100:.2f} %")
     print(f"Annual farm yield : {turbine_results[FV.YLD].sum():.2f} GWh")